# Improve performance by removing unneeded Glob directory cache