@pepibumur 2021-01-19
@adellibovi would you mind updating the CHANGELOG? We can merge afterward.